# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutlineCandidate": {
        "properties": {
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "level": {
            "enum": [
              "volume",
              "chapter"
            ],
            "title": "Level",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "seq": {
            "default": 1,
            "title": "Seq",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "level",
          "title"
        ],
        "title": "OutlineCandidate",
        "type": "object"
      },
      "OutlineExpandRequest": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "count": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "expand_level": {
            "enum": [
              "volume",
              "chapter"
            ],
            "title": "Expand Level",
            "type": "string"
          }
        },
        "required": [
          "expand_level"
        ],
        "title": "OutlineExpandRequest",
        "type": "object"
      },
      "OutlineExpandResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/OutlineCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "expand_level": {
            "enum": [
              "volume",
              "chapter"
            ],
            "title": "Expand Level",
            "type": "string"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "integer"
          },
          "raw_ai_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Ai Output"
          }
        },
        "required": [
          "parent_id",
          "expand_level",
          "candidates"
        ],
        "title": "OutlineExpandResponse",
        "type": "object"
      },
      "OutlineInput": {
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "level": {
            "enum": [
              "total",
              "volume",
              "chapter"
            ],
            "title": "Level",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "seq": {
            "default": 0,
            "title": "Seq",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "level"
        ],
        "title": "OutlineInput",
        "type": "object"
      },
      "OutlineOut": {
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "level": {
            "enum": [
              "total",
              "volume",
              "chapter"
            ],
            "title": "Level",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "seq": {
            "default": 0,
            "title": "Seq",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "level",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "OutlineOut",
        "type": "object"
      },
      "OutlineSummarizeResponse": {
        "description": "「汇总本卷」结果：卷摘要已写回节点 content（带【本卷摘要】标记）。",
        "properties": {
          "outline_id": {
            "title": "Outline Id",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "outline_id",
          "summary"
        ],
        "title": "OutlineSummarizeResponse",
        "type": "object"
      },
      "OutlineUpdate": {
        "description": "PATCH 部分更新。",
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "level": {
            "anyOf": [
              {
                "enum": [
                  "total",
                  "volume",
                  "chapter"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "OutlineUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 小说创作工具",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books/{book}/outlines": {
      "get": {
        "operationId": "list_outlines_api_books__book__outlines_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "in": "query",
            "name": "parent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OutlineOut"
                  },
                  "title": "Response List Outlines Api Books  Book  Outlines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Outlines",
        "tags": [
          "outlines"
        ]
      },
      "post": {
        "operationId": "create_outline_api_books__book__outlines_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlineInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Outline",
        "tags": [
          "outlines"
        ]
      }
    },
...
    "/api/outlines/{outline_id}": {
      "delete": {
        "operationId": "delete_outline_api_outlines__outline_id__delete",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "outline_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Outline Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Outline",
        "tags": [
          "outlines"
        ]
      },
      "patch": {
        "operationId": "update_outline_api_outlines__outline_id__patch",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "outline_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Outline Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlineUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Outline",
        "tags": [
          "outlines"
        ]
      }
    },
    "/api/outlines/{outline_id}/expand": {
      "post": {
        "operationId": "expand_outline_api_outlines__outline_id__expand_post",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "outline_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Outline Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlineExpandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineExpandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Expand Outline",
        "tags": [
          "outlines"
        ]
      }
    },
    "/api/outlines/{outline_id}/summarize-volume": {
      "post": {
        "description": "AI 把本卷各章摘要压成卷摘要并写回节点（记忆金字塔；重跑只替换标记段）。",
        "operationId": "summarize_volume_api_outlines__outline_id__summarize_volume_post",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "outline_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Outline Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineSummarizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summarize Volume",
        "tags": [
          "outlines"
        ]
      }
    },
...
  }
}